Added bools for the otters and changed score UI
NPOs now have a bool that says whether or not they've already been held, and a new variable "otterScore" has been added to keep track of this. When all the otters have been held, the win condition activates. However, this is still under construction, as the triggers sometimes make the counter go up more than once. Will fix this soon.
This change is also reflected in the UI - instead of showing how many shells, the number is how many otters have been held, or "otterScore."