fixed the counter
also commented out the particle stuff you were doing in the scoring section of PlayerController (particle.heldGet). Turns out the error it was throwing was also throwing off the scoring. When you finish the code and have no errors the scoring should still work, but as you work on that don't be worried if the scoring doesn't work! It will work as long as there are no errors with the particles.